requirements.txt
kubernetes==27.2.0
boto3==1.28.69
glueops-helpers @ https://github.com/GlueOps/python-glueops-helpers-library/archive/refs/tags/v0.2.0.zip
PyYAML==6.0.1